Refactor script names
