--- src/f4enix/meshtal_2/main_interface.py
@@ -1,4 +1,0 @@
-import mesh2vtk_2
-
-mesh2vtk = mesh2vtk_2.Interface()
-mesh2vtk.run()
