--- src/f4enix/meshtal_2/mesh2vtk_2/__init__.py
@@ -1,2 +1,0 @@
-from .Modules.FMesh import FMesh
-from .Modules.mesh_parser import CDGSMeshParser, CUVMeshParser, MeshtalParser
