--- src/f4enix/meshtal_2/meshes/cdgs
@@ -1,162 +1,0 @@
-num_meshes     1
-total_source 5.3407E-03
-mesh_id      1
-CDGS produced by D1S-UNED simulation.                   inp=input_2           10/25/24 15:17:08 
-cooling_time 1.000E+00
-mesh_source 5.3407E-03
-energy_type bins
-energy_boundaries    10
- 0.000   0.100   0.200   0.500   0.800   1.000   1.200   2.000
- 5.000  10.000
-mesh_type rec
-mesh_boundaries       5      6      4
-    0.00000E+00   0.00000E+00   0.00000E+00    1.00000E+00   0.00000E+00   0.00000E+00    0.00000E+00   1.00000E+00   0.00000E+00
- 0.00000E+00  7.50000E+00  1.50000E+01  2.25000E+01  3.00000E+01
--5.00000E+01 -3.00000E+01 -1.00000E+01  1.00000E+01  3.00000E+01  5.00000E+01
--5.00000E+01 -1.66667E+01  1.66667E+01  5.00000E+01
-source_data
-      1   4.01260E-08   5.00000E+03    1
- 0       1.00000   4.01260E-08 1
- 4.00589E-08 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00 1.25925E-12 6.57982E-11 0.00000E+00
- 0.00000E+00
-0.875413 0.000000 0.000000 0.000000 0.000000 0.994987 0.969107 0.000000
-0.000000
-      2   3.05556E-08   5.00000E+03    1
- 0       1.00000   3.05556E-08 1
- 3.05159E-08 0.00000E+00 4.07159E-12 0.00000E+00 0.00000E+00 1.76208E-11 1.80227E-11 0.00000E+00
- 0.00000E+00
-0.897565 0.000000 0.994987 0.000000 0.000000 0.920320 0.921364 0.000000
-0.000000
-      3   4.74777E-05   5.00000E+03    1
- 0       1.00000   4.74777E-05 1
- 5.30403E-08 0.00000E+00 0.00000E+00 0.00000E+00 4.74247E-05 0.00000E+00 0.00000E+00 0.00000E+00
- 0.00000E+00
-0.939392 0.000000 0.000000 0.000000 0.994987 0.000000 0.000000 0.000000
-0.000000
-      4   2.64982E-04   5.00000E+03    1
- 0       1.00000   2.64982E-04 1
- 8.58341E-06 0.00000E+00 0.00000E+00 0.00000E+00 2.56399E-04 1.23003E-12 0.00000E+00 0.00000E+00
- 0.00000E+00
-0.989758 0.000000 0.000000 0.000000 0.994987 0.994987 0.000000 0.000000
-0.000000
-      5   6.73699E-08   5.00000E+03    1
- 0       1.00000   6.73699E-08 1
- 6.73551E-08 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00 1.01320E-11 4.71017E-12 0.00000E+00
- 0.00000E+00
-0.621055 0.000000 0.000000 0.000000 0.000000 0.913824 0.994987 0.000000
-0.000000
-      6   2.47577E-09   5.00000E+03    1
- 0       1.00000   2.47577E-09 1
- 2.41526E-09 8.95304E-12 0.00000E+00 0.00000E+00 0.00000E+00 3.29294E-11 1.86230E-11 0.00000E+00
- 0.00000E+00
-0.892756 0.749483 0.000000 0.000000 0.000000 0.895187 0.846038 0.000000
-0.000000
-      7   8.01953E-08   5.00000E+03    1
- 0       1.00000   8.01953E-08 1
- 6.13808E-08 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00 1.40554E-11 1.88004E-08 0.00000E+00
- 0.00000E+00
-0.765673 0.000000 0.000000 0.000000 0.000000 0.870254 0.993587 0.000000
-0.000000
-      8   9.74844E-04   5.00000E+03    1
- 0       1.00000   9.74844E-04 1
- 2.59640E-07 9.32503E-10 0.00000E+00 0.00000E+00 9.74582E-04 1.65150E-09 2.78105E-10 0.00000E+00
- 0.00000E+00
-0.662855 0.994987 0.000000 0.000000 0.994987 0.992858 0.987702 0.000000
-0.000000
-      9   4.05192E-05   5.00000E+03    1
- 0       1.00000   4.05192E-05 1
- 4.04989E-05 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00 2.02414E-08 4.35913E-11 0.00000E+00
- 0.00000E+00
-0.994614 0.000000 0.000000 0.000000 0.000000 0.983976 0.994987 0.000000
-0.000000
-     10   1.29882E-08   5.00000E+03    1
- 0       1.00000   1.29882E-08 1
- 2.22573E-09 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00 1.06391E-08 1.23441E-10 0.00000E+00
- 0.00000E+00
-0.672014 0.000000 0.000000 0.000000 0.000000 0.994682 0.994987 0.000000
-0.000000
-     11   1.21425E-04   5.00000E+03    1
- 0       1.00000   1.21425E-04 1
- 7.73759E-09 0.00000E+00 1.28880E-11 0.00000E+00 1.21417E-04 5.89577E-11 1.08165E-12 0.00000E+00
- 0.00000E+00
-0.826060 0.000000 0.994987 0.000000 0.994987 0.926047 0.994987 0.000000
-0.000000
-     12   4.95729E-08   5.00000E+03    1
- 0       1.00000   4.95729E-08 1
- 3.88030E-09 4.54228E-08 0.00000E+00 0.00000E+00 0.00000E+00 1.46766E-10 1.22944E-10 0.00000E+00
- 0.00000E+00
-0.692486 0.994987 0.000000 0.000000 0.000000 0.797036 0.903487 0.000000
-0.000000
-     13   7.47998E-08   5.00000E+03    1
- 0       1.00000   7.47998E-08 1
- 7.47998E-08 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00
- 0.00000E+00
-0.697398 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
-0.000000
-     14   2.59233E-03   5.00000E+03    1
- 0       1.00000   2.59233E-03 1
- 2.67785E-08 0.00000E+00 0.00000E+00 0.00000E+00 2.04034E-03 6.82555E-09 2.05254E-11 5.51965E-04
- 0.00000E+00
-0.727736 0.000000 0.000000 0.000000 0.709693 0.794215 0.994987 0.994987
-0.000000
-     15   8.88340E-04   5.00000E+03    1
- 0       1.00000   8.88340E-04 1
- 8.57819E-09 0.00000E+00 0.00000E+00 0.00000E+00 4.48342E-04 0.00000E+00 9.30946E-11 4.39990E-04
- 0.00000E+00
-0.915948 0.000000 0.000000 0.000000 0.946469 0.000000 0.994987 0.994987
-0.000000
-     19   4.42494E-10   5.00000E+03    1
- 0       1.00000   4.42494E-10 1
- 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00 4.42494E-10 0.00000E+00
- 0.00000E+00
-0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.994987 0.000000
-0.000000
-     20   7.10761E-08   5.00000E+03    1
- 0       1.00000   7.10761E-08 1
- 6.80637E-08 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00 3.01245E-09 0.00000E+00 0.00000E+00
- 0.00000E+00
-0.689648 0.000000 0.000000 0.000000 0.000000 0.994987 0.000000 0.000000
-0.000000
-     22   1.85286E-08   5.00000E+03    1
- 0       1.00000   1.85286E-08 1
- 1.85272E-08 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00 6.52260E-13 7.01294E-13 0.00000E+00
- 0.00000E+00
-0.994817 0.000000 0.000000 0.000000 0.000000 0.994987 0.994987 0.000000
-0.000000
-     23   9.60119E-08   5.00000E+03    1
- 0       1.00000   9.60119E-08 1
- 8.85887E-08 8.50736E-13 0.00000E+00 0.00000E+00 0.00000E+00 7.42120E-09 1.14996E-12 0.00000E+00
- 0.00000E+00
-0.842249 0.994987 0.000000 0.000000 0.000000 0.994831 0.772091 0.000000
-0.000000
-     24   1.08340E-10   5.00000E+03    1
- 0       1.00000   1.08340E-10 1
- 1.08340E-10 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00
- 0.00000E+00
-0.994987 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
-0.000000
-     25   9.25334E-13   5.00000E+03    1
- 0       1.00000   9.25334E-13 1
- 9.25334E-13 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00
- 0.00000E+00
-0.994987 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
-0.000000
-     27   6.55154E-10   5.00000E+03    1
- 0       1.00000   6.55154E-10 1
- 6.16639E-10 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00 3.85143E-11 0.00000E+00
- 0.00000E+00
-0.876479 0.000000 0.000000 0.000000 0.000000 0.000000 0.994987 0.000000
-0.000000
-     29   4.10095E-04   5.00000E+03    1
- 0       1.00000   4.10095E-04 1
- 1.81223E-08 0.00000E+00 0.00000E+00 0.00000E+00 4.10077E-04 0.00000E+00 0.00000E+00 0.00000E+00
- 0.00000E+00
-0.994987 0.000000 0.000000 0.000000 0.994987 0.000000 0.000000 0.000000
-0.000000
-     30   1.29610E-07   5.00000E+03    1
- 0       1.00000   1.29610E-07 1
- 1.22941E-07 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00 0.00000E+00 6.66969E-09 0.00000E+00
- 0.00000E+00
-0.833994 0.000000 0.000000 0.000000 0.000000 0.000000 0.994987 0.000000
-0.000000
-end_source_data
