--- src/f4enix/meshtal_2/meshes/cuvmsk2
@@ -1,48 +1,0 @@
-d1sune version 412   ld=05/17/23  probid =  04/23/24 00:34:12 
- C                                                                               
- Number of histories used for normalizing tallies =    5000000000.00
-
- Mesh Tally Number      4104
-     1x1x1 cm3 - TBM MESH - NUC. HEAT. DUE TO NEUTRONS                          
- This is a neutron cell-under-voxel mesh tally.
-
- Tally bin boundaries:
-    X direction:    845.00    846.00    847.00    
-	Y direction:    -65.00    -64.00    -63.00    
-	Z direction:    -25.00    -24.00    -23.00   
-	Energy bin number:    1
-    Energy bin boundaries: 1.00E-11 5.00E+01
-
-   INDEX         X         Y         Z    Volume   ncel
-       1   845.500   -64.500   -24.500 1.00000E+00   1
-   1182751 1.00000
- 0.00000E+00
- 0.00000E+00
-       2   845.500   -64.500   -23.500 1.00000E+00   1
-   1182751 1.00000
- 0.00000E+00
- 0.00000E+00
-       3   845.500   -64.500   -22.500 1.00000E+00   1
-   1182751 1.00000
- 0.00000E+00
- 0.00000E+00
-       4   845.500   -64.500   -21.500 1.00000E+00   1
-   1182751 1.00000
- 0.00000E+00
- 0.00000E+00
-       5   845.500   -64.500   -20.500 1.00000E+00   1
-   1182751 1.00000
- 0.00000E+00
- 0.00000E+00
-       6   845.500   -64.500   -19.500 1.00000E+00   1
-   1182751 1.00000
- 0.00000E+00
- 0.00000E+00
-       7   845.500   -64.500   -18.500 1.00000E+00   1
-   1182751 1.00000
- 0.00000E+00
- 0.00000E+00
-       8   845.500   -64.500   -17.500 1.00000E+00   1
-   1182751 1.00000
- 0.00000E+00
- 0.00000E+00
