+++ src/f4enix/output/meshtal/__init__.py
@@ -1,0 +1,4 @@
+from f4enix.output.meshtal.fmesh import Fmesh
+from f4enix.output.meshtal.meshtal import Meshtal
+
+__all__ = ["Fmesh", "Meshtal"]
